+++ app/mdx/mdx-hooks.tsx
@@ -1,10 +1,10 @@
 import { useLoaderData } from "react-router";
 import { SafeMdxRenderer } from "safe-mdx";
 import { mdxParse } from "safe-mdx/parse";
-import type { LoadData, MDXComponents, MdxAttributes } from "./types";
+import type { MDXComponents, MdxAttributes, PostLoaderData } from "./types";
 
-export function useMdxComponent<T extends MDXComponents>(components?: T) {
-  const { attributes, __raw } = useLoaderData<LoadData>();
+export function useMdxComponent(components?: MDXComponents) {
+  const { attributes, __raw } = useLoaderData<PostLoaderData>();
 
   return () => {
     const ast = mdxParse(__raw);
@@ -25,7 +25,7 @@ export const useMdxFiles = () => {
 };
 
 export const useMdxAttributes = () => {
-  const { attributes } = useLoaderData<LoadData>();
+  const { attributes } = useLoaderData<PostLoaderData>();
 
   return attributes;
 };
