+++ app/routes/post.tsx
@@ -3,10 +3,10 @@ import { components } from "~/components/components";
 import Metadata from "~/components/metadata";
 import Metalinks from "~/components/metalinks";
 import tags from "~/components/tags";
-import { useMdxAttributes, useMdxComponent } from "~/lib/mdx-hooks";
-import type { PostLoaderData } from "~/lib/types";
-import { getMetaData, processArticleDate } from "~/utils/posts";
-import { loadMdxRuntime } from "../lib/mdx-runtime";
+import { useMdxAttributes, useMdxComponent } from "~/mdx/mdx-hooks";
+import type { PostLoaderData } from "~/mdx/types";
+import { processArticleDate } from "~/utils/posts";
+import { loadMdxRuntime } from "../mdx/mdx-runtime";
 import type { Route } from "./+types/post";
 
 export async function loader({
@@ -28,11 +28,7 @@ export function meta({ data }: Route.MetaArgs) {
 export default function Post() {
   const Component = useMdxComponent(components);
   const { title, data, links, date } = useMdxAttributes();
-
-  const metadata = getMetaData(data);
-  const metalinks = getMetaData(links);
-
-  const { updatedMetadata, isOldArticle } = processArticleDate(metadata, date);
+  const { metadata, isOldArticle } = processArticleDate(data, date);
 
   return (
     <div className="grid gap-y-4">
@@ -49,7 +45,7 @@ export default function Post() {
           {title}
         </h1>
       </div>
-      <Metadata data={updatedMetadata} />
+      <Metadata data={metadata} />
       {isOldArticle ? (
         <p className="rounded-md overflow-hidden border border-orange-200 dark:border-orange-800 text-orange-700 dark:text-orange-300 sm:px-4 py-3 max-w-[65ch]">
           This has not been updated in the last three months, so this
@@ -59,7 +55,7 @@ export default function Post() {
       <div className="prose prose-headings:text-sm prose-sm dark:prose-invert prose-a:text-indigo-600 hover:prose-a:text-indigo-500 dark:prose-a:text-indigo-400 dark:hover:prose-a:text-indigo-300 prose-a:no-underline py-3 sm:px-4 border border-transparent max-w-none [&>*:not(.code)]:max-w-[60ch] [&>.code]:overflow-x-scroll [&>.code]:max-w-[calc(100vw)] [&>.code]:-mx-8 sm:[&>.code]:-mx-12 [&>.code_code>*]:px-8 sm:[&>.code_code>*]:px-12 [&>.code]:border-y [&>.code]:py-4 [&>.code]:dark:border-gray-800 [&>.code]:dark:bg-black [&>.code]:bg-gray-50 [&>.code]:text-gray-400 prose-h1:before:content-['#'] prose-h2:before:content-['##'] prose-h3:before:content-['###'] prose-h4:before:content-['####'] prose-h5:before:content-['#####'] prose-h6:before:content-['######'] prose-headings:before:mr-2 prose-headings:before:tracking-widest prose-headings:before:text-indigo-400 prose-h1:border-b-2 prose-h2:border-b prose-h1:border-indigo-500 prose-h2:border-indigo-500 prose-h2:pb-4 prose-h2:mb-4 prose-h1:mt-16 prose-h2:mt-12 prose-h3:mt-8 prose-h4:mt-4 prose-h5:mt-2 prose-h6:mt-2 prose-a:text-wrap prose-a:break-words prose-a:[word-break:break-word] text-pretty">
         <Component />
       </div>
-      <Metalinks links={metalinks} />
+      <Metalinks links={links} />
     </div>
   );
 }
