+++ tsconfig.cloudflare.json
@@ -6,7 +6,7 @@
     "app/**/.server/**/*",
     "app/**/.client/**/*",
     "workers/**/*",
-    "worker-configuration.d.ts", "app/lib/.server", "app/lib/.client",
+    "worker-configuration.d.ts", "app/mdx/.server", "app/mdx/.client",
   ],
   "compilerOptions": {
     "composite": true,
